+++ src/input.c
@@ -79,7 +79,7 @@ void fixCursorXPosition(void) {
   E.cursorX = min(E.cursorX, max(0, E.lines[E.cursorY].length + (E.mode == NORMAL ? -1 : 0)));
 }
 
-void moveCursorToLine(int lineNumber) {
+void moveCursorToLine(long lineNumber) {
   E.cursorY = clamp(1, lineNumber, E.numlines) - 1;
   E.cursorX = E.highestLastX;
   fixCursorXPosition();
